add date on task item

+++ src/components/TaskItem.vue
@@ -22,6 +22,8 @@
       <input v-else type="text" name="" id="" :placeholder="task.title" v-model="newTitle" class="my-10" />
       <!-- Print old description before edit -->
       <p v-if="!toggle" class="text-gray-700 text-base mb-4">{{ task.description }}</p>
+      <!-- INSERT DATE -->
+      <p v-if="!toggle" class="text-gray-400 text-base mb-4"> Date: {{moment(task.inserted_at).format('llll') }}</p>
       <!-- Recive a new value for description, show a old description on placeholder-->
       <textarea v-if="toggle" name="" id="" cols="20" rows="6" :placeholder="task.description"
         v-model="newDescription"></textarea>
@@ -69,35 +71,62 @@
 import { ref } from 'vue';
 import { useTaskStore } from '../stores/task';
 import { supabase } from '../supabase';
+import Swal from 'sweetalert2'
+
+import moment from "moment";
+
+// const date = moment().startOf('day').fromNow(); 
+// const date = moment().calendar(); 
+
 // const taskStore = useTaskStore();
 
 const emit = defineEmits(["getTasks"])
+
 
 const toggle = ref(false);
 
 const showInp = () => {
   toggle.value = true;
-
   // emit("getTasks")
 }
-
-
 
 const newTitle = ref("");
 const newDescription = ref("");
 
 // Función para borrar la tarea a través de la store. El problema que tendremos aquí (y en NewTask.vue) es que cuando modifiquemos la base de datos los cambios no se verán reflejados en el v-for de Home.vue porque no estamos modificando la variable tasks guardada en Home. Usad el emit para cambiar esto y evitar ningún page refresh.
-const deleteTask = async(id) => {
-    // await taskStore.deleteTask(props.task.id);
+const deleteTask = async (id) => {
+  const { isConfirmed } = await Swal.fire({
+    title: '¿Está seguro?',
+    text: 'Una vez borrado, no se puede recuperar',
+    showDenyButton: true,
+    confirmButtonText: 'Si, estoy seguro'
+  })
+
+  if (isConfirmed) {
+    new Swal({
+      title: 'Espere por favor',
+      allowOutsideClick: false
+    })
+    Swal.showLoading()
+    // await this.deleteEntry( this.entry.id )
+    // this.$router.push({ name: 'no-entry' })
     await useTaskStore().deleteTask(id);
+    emit("getTasks")
+
+
+    Swal.fire('Eliminado', '', 'success')
+  }
+  // await taskStore.deleteTask(props.task.id);
   // await useTaskStore().fetchTasks();
-  emit("getTasks")
-  alert("Tarea eliminada con exito");
+  // alert("Tarea eliminada con exito");
 };
 
 const toggleTask = async(id,bool) => {
   await useTaskStore().toggleTask(id,bool);
   await useTaskStore().fetchTasks();
+// Falta comprobar el false is_COMPLETE
+    Swal.fire('Tarea Completada','','success')
+ 
   emit("getTasks")
   // alert("Tarea completada con exito");
 }
@@ -118,45 +147,3 @@ const props = defineProps({
 </script>
 
 <style></style>
-
-<!--
-**Hints**
-1. ref() or reactive() can be used here to store the following, think if you want to store them either individually or
-like an object, up to you.
-
-2. Data properties need here are the following: a boolean to store a false**Important variable, a string to store an error,
-a string to store the value of the task that the user can edit, an initial false boolean to hide the inputFIeld used to edit
-the new task detail or details[this is in reference of the task title and the task description].
-
-3. Store the custom emit events that will be used to call the functions of the homeView for editing, deleting and toggling the
-status[completed, not complted] of the taskItem.
-
-4. Function to handle the error with the data properties used to control the error and the the boolean controlling the boolean
-empty variable.
-
-5. Function to handle the edit dialogue where the inputField is displayed and the string used to store the value of the
-inputField will be used here to save the value as a prop on this function.
-
-6. Function to emmit a custom event emit() that takes 2 parameters a name for the custom event and the value that will be
-send via the prop to the parent component. This function can control the toggle completion of the task on the homeview.
-
-7. Function to edit the task information that you decided that the user can edit. This function's body takes in a conditional
-that first checks if the value of the task [either title and description or just title] is empty which if true it runs the
-function used to handle the error on hint4. Else, the conditional sets the first mentioned boolean data property on hint2
-back to its inital boolean value to hide the error message and repeat the same for the data property mentioned 4th on hint2;
-a constant that stores an object that holds the oldValue from the prop item and the value of the task coming from the data
-property mentioned 3rd on hint2; a custom event emit() that takes 2 parameters a name for the custom event and the value
-from the object used on this part of the conditional and lastly this part of the conditional sets the value of input field
-to an empty string to clear it from the ui.
-
-8. Function to emmit a custom event emit() that takes 2 parameters a name for the custom event and the value that will be
-send via the prop to the parent component. This function can control the removal of  the task on the homeview.
--->
-
-<!-- <div class="container">
-    <h3 >{{task.title}}</h3>
-    <button @click="deleteTask">Delete {{task.title}}</button>
-</div> -->
-
-
-
